fix: move methods that are resposible to comsume rabbitmq message
